6 Size

+++ 0x00-hello_world/6-size.c
@@ -10,10 +10,10 @@ int x;
 long int y;
 long long int z;
 float f;
-printf("Size of a char: %b byte(s)\n", sizeof(w));
-printf("Size of an int: %b byte(s)\n", sizeof(x));
-printf("Size of a long int: %b byte(s)\n", sizeof(y));
-printf("Size of a long long int: %b byte(s)\n", sizeof(z));
-printf("Size of a float: %b byte(s)\n", sizeof(f));
+printf("Size of a char: %b byte(s)\n", (unsigned long)sizeof(w));
+printf("Size of an int: %b byte(s)\n", (unsigned long)sizeof(x));
+printf("Size of a long int: %b byte(s)\n", (unsigned long)sizeof(y));
+printf("Size of a long long int: %b byte(s)\n", (unsigned long)sizeof(z));
+printf("Size of a float: %b byte(s)\n", (unsigned long)sizeof(f));
 return (0);
 }
